chore: DashboardProvider and useDashboardContext
