+++ packages/frontend/src/components/common/Dashboard/DashboardRefreshButton.tsx
@@ -4,7 +4,7 @@ import { IconChevronDown, IconRefresh } from '@tabler/icons-react';
 import { useCallback, useEffect, useState } from 'react';
 import { useDashboardRefresh } from '../../../hooks/dashboard/useDashboardRefresh';
 import useToaster from '../../../hooks/toaster/useToaster';
-import { useDashboardContext } from '../../../providers/DashboardProvider';
+import { useDashboardContext } from '../../../providers/DashboardProvider/useDashboardContext';
 import MantineIcon from '../MantineIcon';
 
 const REFRESH_INTERVAL_OPTIONS = [
