+++ packages/frontend/src/features/dateZoom/components/DateZoomInfoOnTile.tsx
@@ -3,7 +3,7 @@ import { Text, Tooltip } from '@mantine/core';
 import { IconCalendarSearch } from '@tabler/icons-react';
 import { FC } from 'react';
 import MantineIcon from '../../../components/common/MantineIcon';
-import { useDashboardContext } from '../../../providers/DashboardProvider';
+import { useDashboardContext } from '../../../providers/DashboardProvider/useDashboardContext';
 
 type Props = {
     chartUuid: string | null;
